requirements.txt
flask==2.2.3
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.5.0
Werkzeug==2.2.3
opencv-python==4.7.0.72
numpy==1.24.3
scikit-image==0.20.0
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
colorama==0.4.6
markupsafe==2.1.2